Final "Alpha"
First version is done, works fine on all methods.
The convergence is working fast for some functions specially for with line search enabled.
The plotting in utils.py is gpt code and is not final or mine.
Using functions for implementing this is really not fun and I'll implement it with classes soon.